+++ com/learn/java/day4/SimpleInterest.java
@@ -5,13 +5,13 @@ public class SimpleInterest {
 	public static void main(String[] args){
 		Scanner scan=new Scanner(System.in);
 		System.out.print("Principal amount\t");
-		float p=scan.nextFloat();
+		double p=scan.nextDouble();
 		System.out.print("Rate of Interest\t");
-		float r=scan.nextFloat();
+		double r=scan.nextDouble();
 		System.out.print("Number of days after receiving loan(From current date)\t");
-		float t=scan.nextFloat();
-		float si=p*r/100*t/365;
-		float net=p+si;
+		double t=scan.nextDouble();
+		double si=p*r/100*t/365;
+		double net=p+si;
 		System.out.print("Simple interest\t: "+si+"\nNet payable amount\t: Rs"+net);
 	
 	}
